requirements.txt
Pillow
pytesseract
opencv-python
pika
requests
python-dotenv